Move layoutButton and layoutChooser to component file
