+++ src/components/layoutButton/index.js
@@ -1,0 +1,1 @@
+export { LayoutButton } from './layoutButton.js';
