+++ src/components/layoutButton/layoutButton.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import LayoutChooser from './LayoutChooser';
-import ToolbarButton from '../viewer/ToolbarButton';
+import { LayoutChooser } from './layoutChooser';
+import ToolbarButton from './../../viewer/ToolbarButton';
 
 export class LayoutButton extends PureComponent {
   static defaultProps = {
